+++ qunit/tests_formatvalidate.js
@@ -143,3 +143,9 @@ test("a13x3y isValid(\"some_body@mail.com\", {alias:\"email\"}", function () {
     var isValid = inputmask.isValid("some_body@mail.com", { alias: "email" });
     equal(isValid, true, "Result " + isValid);
 });
+
+module("Value unmasking");
+test("inputmask.unmask(\"23/03/1973\", { alias: \"dd/mm/yyyy\" })", function () {
+    var unmasked = inputmask.unmask("23/03/1973", { alias: "dd/mm/yyyy" });
+    equal(unmasked, "23031973", "Result " + unmasked);
+});